refactor: :recycle: extract order-metadata-manager
